cpu_memory/image-docker/requirements.txt
Flask
boto3
Pillow
gunicorn
